+++ src/com/anji/integration/Activator.java
@@ -22,44 +22,44 @@ package com.anji.integration;
 import com.anji.util.XmlPersistable;
 
 /**
- * Although any object taking and returning float arrays can implement this, it is meant to be
+ * Although any object taking and returning double arrays can implement this, it is meant to be
  * a simple neural net interface where the arrays are input and output activation patterns.
  * @author Philip Tucker
  */
 public interface Activator extends XmlPersistable {
 
 /**
- * @return Object output array of type float with dimensions dependent on
+ * @return Object output array of type double with dimensions dependent on
  * implementation, given last provided input activation via
- * <code>nextSequence(float[])</code> or <code>nextSequence(float[][])</code>.
- * @see Activator#nextSequence(float[])
- * @see Activator#nextSequence(float[][])
+ * <code>nextSequence(double[])</code> or <code>nextSequence(double[][])</code>.
+ * @see Activator#nextSequence(double[])
+ * @see Activator#nextSequence(double[][])
  */
 public Object next();
 
 /**
  * @param stimuli
- * @return float[] output array given input <code>stimuli</code>.
+ * @return double[] output array given input <code>stimuli</code>.
  */
-public float[] next( float[] stimuli );
+public double[] next( double[] stimuli );
 
 /**
  * @param stimuli
- * @return float[][] sequence of output arrays given input sequence <code>stimult</code>.
+ * @return double[][] sequence of output arrays given input sequence <code>stimult</code>.
  */
-public float[][] nextSequence( float[][] stimuli );
+public double[][] nextSequence( double[][] stimuli );
 
 /**
  * @param stimuli
- * @return float[][] output array given input <code>stimuli</code>.
+ * @return double[][] output array given input <code>stimuli</code>.
  */
-public float[][] next( float[][] stimuli );
+public double[][] next( double[][] stimuli );
 
 /**
  * @param stimuli
- * @return float[][][] sequence of output arrays given input sequence <code>stimult</code>.
+ * @return double[][][] sequence of output arrays given input sequence <code>stimult</code>.
  */
-public float[][][] nextSequence( float[][][] stimuli );
+public double[][][] nextSequence( double[][][] stimuli );
 
 /**
  * @return String XML representation of object.
@@ -79,12 +79,12 @@ public String getName();
 /**
  * @return min response value
  */
-public float getMinResponse();
+public double getMinResponse();
 
 /**
  * @return max response value
  */
-public float getMaxResponse();
+public double getMaxResponse();
 
 /**
  * @return dimension(s) of input array
